requirements.txt
acres==0.2.0
certifi==2024.12.14
charset-normalizer==3.4.1
ci-info==0.3.0
click==8.1.8
configobj==5.0.9
configparser==7.1.0
et-xmlfile==1.1.0
etelemetry==0.3.1
filelock==3.17.0
fitz==0.0.1.dev2
httplib2==0.22.0
idna==3.10
importlib_resources==6.5.2
isodate==0.6.1
looseversion==1.3.0
lxml==5.3.0
networkx==3.2.1
nibabel==5.3.2
nipype==1.9.2
numpy==2.0.2
openpyxl==3.1.5
packaging==24.2
pandas==2.2.3
pathlib==1.0.1
pdf2image==1.17.0
pillow==11.1.0
prov==2.0.1
puremagic==1.28
pydot==3.0.4
PyMuPDF==1.25.2
pyparsing==3.2.1
pytesseract==0.3.13
python-dateutil==2.9.0.post0
pytz==2024.2
pyxnat==1.6.2
rdflib==6.3.2
requests==2.32.3
scipy==1.13.1
simplejson==3.19.3
six==1.16.0
traits==7.0.2
typing_extensions==4.12.2
tzdata==2024.2
urllib3==2.3.0
zipp==3.21.0
